Move plugins helpers into their own file
