+++ packages/plugins/build-report/src/esbuild.ts
@@ -2,7 +2,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { getAbsolutePath, getEsbuildEntries, isInjectionFile } from '@dd/core/helpers';
+import { isInjectionFile } from '@dd/core/helpers/plugins';
+import { getAbsolutePath, getEsbuildEntries } from '@dd/core/helpers';
 import type {
     Logger,
     Entry,
