+++ packages/plugins/build-report/src/xpack.ts
@@ -2,7 +2,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { getAbsolutePath, isInjectionFile } from '@dd/core/helpers';
+import { isInjectionFile } from '@dd/core/helpers/plugins';
+import { getAbsolutePath } from '@dd/core/helpers';
 import type {
     Logger,
     Entry,
