+++ packages/tests/src/unit/plugins/telemetry/index.test.ts
@@ -2,7 +2,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { debugFilesPlugins } from '@dd/core/helpers';
+import { debugFilesPlugins } from '@dd/core/helpers/plugins';
 import type { GlobalContext, Options } from '@dd/core/types';
 import { addMetrics } from '@dd/telemetry-plugin/common/aggregator';
 import type { MetricToSend } from '@dd/telemetry-plugin/types';
